Project Attribute Modify * 경고를 에러로 처리 * 경고수준 3->4 header.h * define 왠지 위에 있어야 할 것 같아서 윗 줄로 이동 Ship.cpp * PrintPosition() 함수에서 for문의 i를 int로 자료형 변경

+++ ConsoleApplication3/Ship.cpp
@@ -89,7 +89,7 @@ HitResult CShip::HitCheck(Position hitPos)
 
 void CShip::PrintPosition()
 {
-	for (size_t i = 0; i < m_Pos.size(); i++)
+	for (int i = 0; i < m_Pos.size(); i++)
 	{
 		printf("%s Position[%d] - %c%c\n", m_Name.c_str(), i, m_Pos[i].x, m_Pos[i].y);
 	}
